+++ examples/libhello/src/test/include/ExternalTypes.h
@@ -105,6 +105,26 @@ public:
 
     virtual void some_Method( uint8_t ) = 0;
 };
+
+class ClassWithOverloads {
+public:
+    struct StructWithOverloads {
+        virtual std::string overloadedMethod( );
+        virtual std::string overloadedMethod( const std::string& input );
+
+        void overloadedAccessors( int32_t value );
+        int32_t overloadedAccessors( ) const;
+    private:
+        int32_t m_someField;
+    };
+
+    virtual std::string oneOverloadNotExposed( ) = 0;
+    virtual std::string oneOverloadNotExposed( const std::string& ) = 0;
+
+    virtual std::string allOverloadsExposed( const std::string& ) = 0;
+    virtual std::string allOverloadsExposed( const std::vector<std::string>& ) = 0;
+};
+
 }  // namespace external
 
 namespace test