+++ examples/libhello/src/test/src/ExternalTypes.cpp
@@ -67,4 +67,27 @@ ExternalStruct::_set_aStruct( const even_more_external::AlienStructure& value )
     m_some_struct = value;
 }
 
+std::string
+ClassWithOverloads::StructWithOverloads::overloadedMethod( )
+{
+    return {};
+}
+
+std::string
+ClassWithOverloads::StructWithOverloads::overloadedMethod( const std::string& input )
+{
+    return input;
+}
+
+void
+ClassWithOverloads::StructWithOverloads::overloadedAccessors( int32_t value ) {
+    m_someField = value;
+}
+
+int32_t
+ClassWithOverloads::StructWithOverloads::overloadedAccessors( ) const
+{
+    return m_someField;
+}
+
 }  // namespace external